third_party/WebKit/LayoutTests/editing/execCommand/indent-empty-root.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>|</div>',
        'indent',
        [
            '<div contenteditable>',
                '<blockquote style="margin: 0 0 0 40px; border: none; padding: 0px;">',
                    '|<br>',
                '</blockquote>',
            '</div>',
        ].join('')),
        'Indent in an empty editable element');
</script>
